requirements.txt
rdflib~=7.1.3
requests~=2.32.3
SPARQLWrapper~=2.0.0
python-slugify~=8.0.4